requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
click==8.1.7
Django==5.1
django-auditlog==3.0.0
django-environ==0.11.2
django-hijack==3.6.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
unpoly==3.1.0
